Rename package ikfastpy → ssik
The ikfastpy name misrepresents the new architecture once IKFast is no longer
the core engine. The rebuild (see #37) routes chains through IK-Geo
subproblems, with specialist solvers registered via entry-points; the LGPL
ikfast tree stays quarantined in _vendor/ as a Pieper-class reference.

- Package moved to src/ssik/ via git mv.
- Imports rewritten across src and tests; pyproject.toml, ruff, mypy configs
  updated; URDF loader error messages, vendored logger names ('ssik.ikfast'),
  and docs reference the new name. GitHub repo URLs unchanged (only the
  PyPI/import name is renamed).
- packaging/ikfastpy-shim/ contains a separate ikfastpy package that
  re-exports ssik with a DeprecationWarning. Ships at v0.1 as a one-release
  migration shim; verified the shim imports cleanly and warns once.
- uv.lock regenerated under the new project name; version still VCS-driven.

ruff + ruff-format + mypy + fast pytest all green. mkdocs --strict builds.
Slow suite (Puma 560 Transform6D, UR5 slow) untouched.

Part of #37. Fixes #38.
