+++ .gitignore
@@ -13,7 +13,7 @@ venv/
 env/
 
 # Generated by hatch-vcs
-src/ikfastpy/_version.py
+src/ssik/_version.py
 
 # Build artefacts
 *.so
@@ -27,7 +27,7 @@ CMakeFiles/
 cmake_install.cmake
 Makefile
 
-# Generated solvers (development; the package will use ~/.cache/ikfastpy/ at runtime)
+# Generated solvers (development; the package will use ~/.cache/ssik/ at runtime)
 generated/
 
 # Tests
