+++ packaging/ikfastpy-shim/pyproject.toml
@@ -1,0 +1,29 @@
+[build-system]
+requires = ["hatchling>=1.27"]
+build-backend = "hatchling.build"
+
+[project]
+name = "ikfastpy"
+version = "0.0.1.post0"
+description = "Deprecation shim: the 'ikfastpy' package has been renamed to 'ssik'."
+readme = "README.md"
+license = { text = "LGPL-3.0-or-later" }
+requires-python = ">=3.11"
+authors = [{ name = "Siddhartha Srinivasa" }]
+keywords = ["robotics", "inverse-kinematics", "ikfast", "deprecated"]
+classifiers = [
+    "Development Status :: 7 - Inactive",
+    "Intended Audience :: Science/Research",
+    "Programming Language :: Python :: 3",
+    "Topic :: Scientific/Engineering",
+]
+dependencies = [
+    "ssik>=0.1.0",
+]
+
+[project.urls]
+Repository = "https://github.com/siddhss5/ikfastpy"
+Issues = "https://github.com/siddhss5/ikfastpy/issues"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/ikfastpy"]
