--- src/ikfastpy/__init__.py
@@ -1,5 +1,0 @@
-"""ikfastpy — standalone Python implementation of the IKFast analytic IK solver."""
-
-from ikfastpy._version import __version__
-
-__all__ = ["__version__"]
