+++ src/ssik/__init__.py
@@ -1,0 +1,5 @@
+"""ssik — pluggable analytical inverse-kinematics library for Python."""
+
+from ssik._version import __version__
+
+__all__ = ["__version__"]
